move example to root bindings dir
